'fix_jackpot'

+++ src/main/java/bet/lucky/game/external_dto/response/JackpotResponse.java
@@ -4,10 +4,11 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 public class JackpotResponse implements Serializable {
     private static final long serialVersionUID = 7753109864935849655L;
-    Double jackpot;
+    BigDecimal jackpot;
 }
